fix(paths): improve getAllPaths performance and readability
